refactor trendline file
